refactor(services): split service boundaries, inject config deps
- Extract OrdersService from PortfolioService: create_order,
  list_local_orders, cancel_order, fill_order moved to dedicated class;
  order router endpoints now depend on OrdersService directly
- Inject ConfigRepository into PortfolioService via __init__ instead of
  constructing it inline at 4 call sites; dependencies.py passes the
  singleton get_config_repo() result
- Remove orders_repo from PortfolioService (no longer owns order logic)
