+++ api/services/portfolio_service.py
@@ -12,11 +12,8 @@ from fastapi import HTTPException
 
 from api.models.portfolio import (
     ConcentrationGroup,
-    CreateOrderRequest,
     CreatePositionRequest,
     EarningsProximityResponse,
-    FillOrderRequest,
-    FillOrderResponse,
     PartialCloseEvent,
     PartialCloseRequest,
     Position,
@@ -30,7 +27,6 @@ from api.models.portfolio import (
     ClosePositionRequest,
 )
 from api.repositories.config_repo import ConfigRepository
-from api.repositories.orders_repo import OrdersRepository
 from api.repositories.positions_repo import PositionsRepository
 from api.utils.files import get_today_str
 from swing_screener.portfolio.state import (
@@ -164,8 +160,8 @@ def _round_price(value: float) -> float:
     return float(Decimal(str(value)).quantize(Decimal("0.01"), rounding=ROUND_HALF_UP))
 
 
-def _manage_cfg_from_app() -> ManageStateConfig:
-    manage = ConfigRepository().get().manage
+def _manage_cfg_from_repo(config_repo: ConfigRepository) -> ManageStateConfig:
+    manage = config_repo.get().manage
     return ManageStateConfig(
         breakeven_at_R=manage.breakeven_at_r,
         trail_sma=manage.trail_sma,
@@ -234,12 +230,12 @@ class PortfolioService:
     def __init__(
         self,
         positions_repo: PositionsRepository,
-        orders_repo: Optional[OrdersRepository] = None,
-        provider: Optional[MarketDataProvider] = None
+        provider: Optional[MarketDataProvider] = None,
+        config_repo: Optional[ConfigRepository] = None,
     ) -> None:
         self._positions_repo = positions_repo
-        self._orders_repo = orders_repo
         self._provider = provider or get_default_provider()
+        self._config_repo = config_repo or ConfigRepository()
 
     def _fetch_last_prices(self, tickers: list[str]) -> dict[str, float]:
         if not tickers:
@@ -397,7 +393,7 @@ class PortfolioService:
             for position in positions
         )
         eurusd_rate = self._eurusd_rate() if has_usd_positions else 1.0
-        account_currency = getattr(ConfigRepository().get().risk, "account_currency", "EUR")
+        account_currency = getattr(self._config_repo.get().risk, "account_currency", "EUR")
 
         positions_with_metrics = [
             self._build_position_with_metrics(
@@ -458,7 +454,7 @@ class PortfolioService:
             total_shares = sum(e.shares_closed for e in partial_close_events)
             blended_r = sum(e.shares_closed * e.r_at_close for e in partial_close_events) / total_shares
 
-        account_currency = getattr(ConfigRepository().get().risk, "account_currency", "EUR")
+        account_currency = getattr(self._config_repo.get().risk, "account_currency", "EUR")
         position_currency = detect_currency(ticker)
         r_fx_adjusted: Optional[float] = None
         entry_fx_rate_raw = position.get("entry_fx_rate")
@@ -630,7 +626,7 @@ class PortfolioService:
             country_risk[country] = country_risk.get(country, 0.0) + position.total_risk
             country_count[country] = country_count.get(country, 0) + 1
 
-        threshold = float(getattr(ConfigRepository().get().risk, "max_concentration_pct", 60.0))
+        threshold = float(getattr(self._config_repo.get().risk, "max_concentration_pct", 60.0))
         groups: list[ConcentrationGroup] = []
         for country, risk_amount in sorted(country_risk.items(), key=lambda item: item[1], reverse=True):
             risk_pct = (risk_amount / open_risk * 100.0) if open_risk > 0 else 0.0
@@ -685,147 +681,6 @@ class PortfolioService:
 
         _earnings_cache[normalized_ticker] = (today, result)
         return result
-
-    def create_order(self, request: CreateOrderRequest) -> dict:
-        """Create a pending entry order in orders.json."""
-        if self._orders_repo is None:
-            raise HTTPException(status_code=503, detail="Orders repository not configured")
-
-        ticker = request.ticker.upper()
-        orders, _ = self._orders_repo.list_orders()
-
-        if request.order_kind == "entry":
-            pending_entry = any(
-                o.get("ticker") == ticker
-                and o.get("status") == "pending"
-                and o.get("order_kind") == "entry"
-                for o in orders
-            )
-            if pending_entry:
-                raise HTTPException(status_code=409, detail=f"{ticker}: pending entry order already exists.")
-
-            positions, _ = self._positions_repo.list_positions(status="open")
-            open_position = next((p for p in positions if p.get("ticker") == ticker), None)
-
-            if request.entry_mode == "ADD_ON":
-                if not open_position:
-                    raise HTTPException(status_code=409, detail=f"{ticker}: no open position found for add-on order.")
-            elif open_position:
-                raise HTTPException(
-                    status_code=409,
-                    detail=f"{ticker}: open position already exists. Create this as an ADD_ON order instead.",
-                )
-
-        existing_ids = {o.get("order_id", "") for o in orders}
-        base = f"ORD-{ticker}"
-        n = 1
-        order_id = f"{base}-{n:03d}"
-        while order_id in existing_ids:
-            n += 1
-            order_id = f"{base}-{n:03d}"
-
-        isin = request.isin or _resolve_isin(ticker)
-        order = {
-            "order_id": order_id,
-            "ticker": ticker,
-            "status": "pending",
-            "order_type": request.order_type,
-            "quantity": request.quantity,
-            "limit_price": request.limit_price,
-            "stop_price": request.stop_price,
-            "order_date": get_today_str(),
-            "filled_date": None,
-            "entry_price": None,
-            "notes": request.notes.strip(),
-            "order_kind": request.order_kind,
-            "parent_order_id": None,
-            "position_id": request.position_id if request.entry_mode == "ADD_ON" else None,
-            "tif": "GTC",
-            "fee_eur": None,
-            "fill_fx_rate": None,
-            "isin": isin,
-            "thesis": request.thesis,
-        }
-        self._orders_repo.append_order(order)
-        return order
-
-    def list_local_orders(self, status: Optional[str] = None) -> dict:
-        """List locally stored orders from orders.json."""
-        if self._orders_repo is None:
-            raise HTTPException(status_code=503, detail="Orders repository not configured")
-        orders, asof = self._orders_repo.list_orders(status=status)
-        return {"orders": orders, "asof": asof}
-
-    def cancel_order(self, order_id: str) -> dict:
-        """Cancel a pending local order."""
-        if self._orders_repo is None:
-            raise HTTPException(status_code=503, detail="Orders repository not configured")
-
-        order = self._orders_repo.cancel_order(order_id)
-        if order is None:
-            raise HTTPException(status_code=404, detail=f"Order {order_id} not found")
-        if order.get("status") != "cancelled":
-            raise HTTPException(status_code=409, detail=f"Order {order_id} is already {order.get('status')}")
-
-        return {"order_id": order_id, "status": "cancelled"}
-
-    def fill_order(self, order_id: str, request: FillOrderRequest) -> FillOrderResponse:
-        """Mark a pending order as filled and create the open position."""
-        if self._orders_repo is None:
-            raise HTTPException(status_code=503, detail="Orders repository not configured")
-
-        order = self._orders_repo.get_order(order_id)
-        if order is None:
-            raise HTTPException(status_code=404, detail=f"Order {order_id} not found")
-        if order.get("status") != "pending":
-            raise HTTPException(status_code=409, detail=f"Order {order_id} is already {order.get('status')}")
-
-        ticker = order["ticker"]
-        stop_price = request.stop_price if request.stop_price is not None else order.get("stop_price")
-        if not stop_price or stop_price <= 0:
-            raise HTTPException(status_code=422, detail=f"No valid stop price for order {order_id}")
-        if stop_price >= request.filled_price:
-            raise HTTPException(status_code=422, detail="stop_price must be below filled_price")
-
-        updates = {
-            "status": "filled",
-            "entry_price": request.filled_price,
-            "filled_date": request.filled_date,
-            "fee_eur": request.fee_eur,
-            "fill_fx_rate": request.fill_fx_rate,
-            "stop_price": stop_price,
-        }
-        self._orders_repo.update_order(order_id, updates)
-
-        isin = order.get("isin") or _resolve_isin(ticker)
-        position_id = f"POS-{uuid.uuid4().hex[:8].upper()}"
-        initial_risk = round(request.filled_price - stop_price, 4)
-
-        new_position: dict = {
-            "position_id": position_id,
-            "ticker": ticker,
-            "status": "open",
-            "entry_date": request.filled_date,
-            "entry_price": request.filled_price,
-            "stop_price": stop_price,
-            "shares": order["quantity"],
-            "initial_risk": initial_risk,
-            "source_order_id": order_id,
-            "isin": isin,
-            "thesis": order.get("thesis"),
-            "notes": order.get("notes", ""),
-            "entry_fee_eur": request.fee_eur,
-            "entry_fx_rate": request.fill_fx_rate,
-        }
-
-        data = self._positions_repo.read()
-        positions = data.get("positions", [])
-        positions.append(new_position)
-        data["positions"] = positions
-        data["asof"] = get_today_str()
-        self._positions_repo.write(data)
-
-        return FillOrderResponse(order_id=order_id, position=Position(**new_position))
 
     def create_position(self, request: CreatePositionRequest) -> Position:
         """Register a position directly (after manual fill at DeGiro)."""
@@ -1018,7 +873,7 @@ class PortfolioService:
 
     def _resolve_manage_cfg(self, payload: Optional[dict] = None) -> ManageStateConfig:
         if payload is None:
-            return _manage_cfg_from_app()
+            return _manage_cfg_from_repo(self._config_repo)
         return ManageStateConfig(
             breakeven_at_R=float(payload.get("breakeven_at_r", 1.0)),
             trail_sma=int(payload.get("trail_sma", 20)),
